+++ src/App.tsx
@@ -4,7 +4,7 @@ import { withAuthenticator, AmplifySignOut } from '@aws-amplify/ui-react';
 import { Amplify, Auth } from 'aws-amplify';
 import { Home } from './views/Home';
 import { Team as TeamView } from './views/Team';
-import { Player as PlayerView } from './views/Player';
+import { Player as PlayerView } from './views/Player/Player';
 import { Admin } from './views/Admin';
 import awsconfig from './aws-exports';
