+++ src/views/Player/Player.styles.tsx
